+++ SpaceGamePrelim/Components/ZoomMovementComp.cpp
@@ -1,8 +1,6 @@
 #include "ZoomMovementComp.h"
 #include "../Frame/ZoomManager.h"
 
-#include <iostream>
-
 
 ZoomMovementComp::ZoomMovementComp()
 {
@@ -23,22 +21,16 @@ ZoomMovementComp::~ZoomMovementComp()
 //
 void ZoomMovementComp::Execute()
 {
-
 	Vector CurrentVelocity = m_Owner->GetVelocity();
 
 	// Normalize the set velocity by the current zoom scale
 	Vector NewVelocity(
-		CurrentVelocity.getX() * ZoomManager::Instance()->GetScale(),
-		CurrentVelocity.getY() * ZoomManager::Instance()->GetScale());
+		CurrentVelocity.getX() * (float)ZoomManager::Instance()->GetScale(),
+		CurrentVelocity.getY() * (float)ZoomManager::Instance()->GetScale());
 
 	// Trying to eliminate garbage, small numbers from the floats
 	if (abs(NewVelocity.getX()) < 0.00001) NewVelocity.setX(0.0);
 	if (abs(NewVelocity.getY()) < 0.00001) NewVelocity.setY(0.0);
-
-
-	//std::cout << "Old velocity " << CurrentVelocity.getX() << " " << CurrentVelocity.getY() << std::endl;
-	//std::cout << "New velocity " << NewVelocity.getX() << " " << NewVelocity.getY() << std::endl;
-
 
 	m_Owner->SetVelocity(NewVelocity);
 }
